requirements.txt
absl-py==1.3.0
ale-py==0.8.1
alembic==1.8.1
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
astunparse==1.6.3
async-generator==1.10
attrs==22.1.0
AutoROM==0.4.2
AutoROM.accept-rom-license==0.6.1
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.24
certipy==0.1.3
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1631636250774/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093492116/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1626371162869/work
click==8.1.7
cloudpickle==3.0.0
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
conda==4.10.3
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1618231390031/work
contourpy==1.0.6
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1633983255347/work
cycler==0.11.0
debugpy==1.6.3
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
etils==0.9.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.9.0
flatbuffers==22.10.26
fonttools==4.38.0
fsspec==2023.4.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.57.0
greenlet==2.0.1
grpcio==1.50.0
gym==0.26.2
gym-notices==0.0.8
h5py==3.7.0
hydra-core==1.3.2
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1609836280497/work
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==5.0.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
isodate==0.6.1
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.0
jupyter-resource-usage==0.6.3
jupyter-server==1.23.1
jupyter-telemetry==0.1.0
jupyter_client==7.4.4
jupyter_core==5.0.0
jupyterhub==1.5.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.16.2
keras==2.11.0
kiwisolver==1.4.4
libclang==14.0.6
lz4==4.3.2
Mako==1.2.3
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba_1632770295204/work
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
moviepy==1.0.3
mpmath==1.3.0
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nbgitpuller==1.1.1
nest-asyncio==1.5.6
networkx==3.0
notebook==6.5.2
notebook_shim==0.2.2
nteract-on-jupyter==2.1.3
numpy==1.23.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==21.3
pamela==1.0.0
pandas==1.5.1
pandas-datacube==0.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
proglog==0.1.10
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.19.6
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1610094799048/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
Pygments==2.13.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1633192417276/work
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291451001/work
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.6
PyYAML==6.0.1
pyzmq==24.0.1
rdflib==6.2.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1626393743643/work
requests-oauthlib==1.3.1
rsa==4.9
ruamel-yaml-conda @ file:///home/conda/feedstock_root/build_artifacts/ruamel_yaml_1611943432947/work
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
Send2Trash==1.8.0
sentry-sdk==1.39.0
setproctitle==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SPARQLWrapper==2.0.0
SQLAlchemy==1.4.43
stack-data==0.6.0
sympy==1.12
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensordict==0.2.1
tensorflow==2.11.0
tensorflow-datasets==4.7.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.27.0
tensorflow-metadata==1.11.0
termcolor==2.1.1
terminado==0.17.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchrl==0.2.1
torchvision==0.16.1+cu118
tornado==6.2
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1632160078689/work
traitlets==5.5.0
triton==2.1.0
typing_extensions==4.4.0
urllib3==1.26.18
wandb==0.16.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==3.6.1
wrapt==1.14.1
zipp==3.10.0
